Migration to Cicerone. Migration to splash dep (#8)
* Refactor navigation. Migrate to Cicerone. Migrate to splashScreen dep.

+++ data/network/src/main/kotlin/com/grappim/network/model/waybill/GetWaybillByBarcodeRequestDTO.kt
@@ -8,5 +8,5 @@ data class GetWaybillByBarcodeRequestDTO(
     @SerialName("barcode")
     val barcode: String,
     @SerialName("waybillId")
-    val waybillId: Int
+    val waybillId: Long
 )
